Add lowercase formatter in harvester
This adds a new formatter to the harvesting fetch consumer to make
all values lowercase. For example, by convention we want all of our
AGROVOC subjects to be lower case, but we can't always update the
repository metadata ourselves. This allows us to normalize values
upon harvesting without having to create mappings.