docker/angularBuild/Dockerfile: Use Ubuntu 20.04
Use Ubuntu 20.04 LTS in the Angular container instead of 18.04. We
need to do a small workaround to set the timezone because it seems
that the Ubuntu 20.04 container doesn't have this variable set:

> DEBIAN_FRONTEND=noninteractive

Earlier versions had this, which would cause apt to not prompt us
during installation. We could either set that here, or just set the
timezone manually before invoking apt. The default timezone is UTC
anyways.