Use a custom image for backend
The backend needs unoconv to be able to create Word and PDF reports.
Currently the backend API is just using `node:12` directly, and we
have to remember to re-install unoconv manually every time we re-up
the service. This configures the backend API to use a simple image
that uses "FROM node:12" and includes unoconv at build time.