Health check and new way of harvest data
This implements a new method for harvesting DSpace repositories as
well as a "health check" plugin that runs after harvesting to find
duplicate and missing items. The new harvesting method harvests 10
items at a time and is much faster (at least 3x) than the previous
method.

We still need to do some cleanup work on this:

- Commit button should only be active if `openrxv-items-temp` has
data, otherwise if you click it twice it will overwrite the live
`openrxv-items-final` index.
- Some documentation tips in the admin dashboard to help users with
the values that are expected in various new fields
- Making the "DSpace add missing items" plugin hidden, as it is used
internally by the health check plugin and should not be enabled
manually

See: https://github.com/ilri/OpenRXV/pull/96