Adjust backend and frontend .gitignore
We should not ignore package-lock.json because it is used when you
run `npm ci` in a continuous integration environment where you want
to install the *exact* versions last tested. A normal `npm install`
does not consult this file and developers should not keep local ch-
anges there.

Basically, `npm install` is not deterministic: depending on your ~,
^, etc you can get different versions of packages. `npm ci` is the
opposite, in that it only installs the exact versions we have in
the package-lock.json file. This makes it better for CI (and faster).